requirements.txt
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
pandas>=2.1
h5py>=3.10
joblib>=1.3
pyyaml>=6.0

